requirements.txt
Django==5.0.6
djangorestframework==3.15.1
passlib==1.7.4
pillow==10.2.0
PyJWT==2.8.0
python-decouple==3.8